:sparkles: Strava polyline analysis code

+++ pipeline/tableau.py
@@ -12,20 +12,21 @@ log_path = os.path.join(abs_path, "..", "log")
 tableau_path = os.path.join(abs_path, "..", "extracts")
 
 
-def makeConversion(df, tableau_conf):
+def makeConversion(df, file_name):
     """convert dataframe to .hyper extract"""
 
     if not os.path.isdir(tableau_path):
         os.mkdir(tableau_path)
 
     # make conversion
-    df_tmp = pandleau(df)
+    if df.shape[1] > 0:
+        df_tmp = pandleau(df)
 
-    # overwrite with new file
-    file_out = os.path.join(tableau_path, "{file}.hyper".format(file = tableau_conf['FILE']))
-    if os.path.isfile(file_out):
-        os.remove(file_out)
-    df_tmp.to_tableau(file_out, add_index=False)
+        # overwrite with new file
+        file_out = os.path.join(tableau_path, f"{file_name}.hyper")
+        if os.path.isfile(file_out):
+            os.remove(file_out)
+        df_tmp.to_tableau(file_out, add_index=False)
 
 def cleanLogs():
     """clean log files function"""
@@ -42,4 +43,4 @@ def cleanLogs():
     for file in files_logs:
         os.rename(file, os.path.join(log_path, file))
 
-    print("\nDirectory cleaned")
+    print("Directory cleaned")
